+++ packages/startupjs/templates/simple/vite.config.mjs
@@ -1,0 +1,9 @@
+import startupjsPlugin from '@startupjs/vite-plugin-startupjs'
+import { defineConfig } from 'vite'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    startupjsPlugin()
+  ]
+})