+++ packages/ui/components/draggable/index.js
@@ -1,0 +1,3 @@
+export { default as Draggable } from './Draggable'
+export { default as Droppable } from './Droppable'
+export { DragDropProvider } from './DragDropContext'