+++ scripts/babel/createModuleMap.js
@@ -14,6 +14,7 @@ const MODULE_DIRS = {
   '@startupjs/ui': [
     { source: path.join(PACKAGES_DIR, 'ui/components') },
     { source: path.join(PACKAGES_DIR, 'ui/components/toast'), includeFiles: true },
+    { source: path.join(PACKAGES_DIR, 'ui/components/draggable'), includeFiles: true },
     { source: path.join(PACKAGES_DIR, 'ui/hooks'), includeFiles: true },
     { source: path.join(PACKAGES_DIR, 'ui/theming'), includeFiles: true },
     { source: path.join(PACKAGES_DIR, 'ui/uiAppPlugin.js'), includeFiles: true },